move ServiceInfo and ServiceStats to own file
