+++ config.go
@@ -115,27 +115,6 @@ type ServiceConfig struct {
 	DoozerUpdateInterval time.Duration `json:"-"`
 }
 
-type ServiceStatistics struct {
-	Clients        int32
-	StartTime      string
-	LastRequest    string
-	RequestsServed int64
-
-	// For now this will be since startup, we might change it later to be for a given sample interval
-	AverageResponseTime time.Duration
-	TotalDuration       time.Duration `json:"-"`
-}
-
-type ServiceInfo struct {
-	Config     *ServiceConfig
-	Registered bool
-	Stats      ServiceStatistics `json:"-"`
-}
-
-func (s *ServiceInfo) GetConfigPath() string {
-	return "/services/" + s.Config.Name + "/" + s.Config.Version + "/" + s.Config.Region + "/" + s.Config.ServiceAddr.IPAddress + "/" + strconv.Itoa(s.Config.ServiceAddr.Port)
-}
-
 type ClientConfig struct {
 	Log                       Logger        `json:"-"`
 	DoozerConfig              *DoozerConfig `json:"-"`
